+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Prevent duplicate model registration based on task output definition
+
 ## [0.22.0] - 2022-08-09
 
 ### Added